requirements.txt
RPi.GPIO
python3-pygame
